refactor(query): extract relation-ir planner module
