+++ crates/jazz-tools/src/query_manager/graph.rs
@@ -26,15 +26,14 @@ use super::graph_nodes::recursive_relation::{
     CorrelationSource, RecursiveHop, RecursiveRelationNode,
 };
 use super::graph_nodes::select_element::SelectElementNode;
-use super::graph_nodes::sort::{SortDirection, SortNode};
+use super::graph_nodes::sort::SortNode;
 use super::graph_nodes::subgraph::SubgraphTemplate;
 use super::graph_nodes::union::UnionNode;
 use super::graph_nodes::{NodeId, RowNode, SourceContext, SourceNode, TransformNode};
 use super::index::ScanCondition;
-use super::query::{Condition, Conjunction, JoinSpec, Query, RecursiveSpec};
-use super::relation_ir::{
-    JoinKind, OrderDirection, PredicateCmpOp, PredicateExpr, RelExpr, RowIdRef, ValueRef,
-};
+use super::query::{Condition, Query};
+use super::relation_ir::RelExpr;
+use super::relation_ir_query_plan::{lower_query_relation_ir, lower_relation_to_execution_query};
 use super::session::Session;
 use super::types::{
     ColumnDescriptor, ColumnName, ColumnType, ComposedBranchName, Row, RowDelta, RowDescriptor,
@@ -95,460 +94,6 @@ pub struct QueryGraph {
     pub table_descriptors: Vec<RowDescriptor>,
     /// Combined descriptor for output (all columns from all tables).
     pub combined_descriptor: RowDescriptor,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct QueryEnvelope<'a> {
-    core: &'a RelExpr,
-    order_by: Vec<(String, SortDirection)>,
-    offset: usize,
-    limit: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct LinearJoinInfo {
-    base_table: TableName,
-    base_scope: String,
-    conditions: Vec<Condition>,
-    joins: Vec<JoinSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct RuntimeCorePlan {
-    table: TableName,
-    conditions: Vec<Condition>,
-    joins: Vec<JoinSpec>,
-    result_element_index: Option<usize>,
-    recursive: Option<RecursiveSpec>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct GatherJoinInfo {
-    plan: RuntimeCorePlan,
-    current_scope: String,
-}
-
-fn to_runtime_column(column: &str) -> String {
-    if column == "id" {
-        "_id".to_string()
-    } else {
-        column.to_string()
-    }
-}
-
-fn flatten_predicate_terms<'a>(predicate: &'a PredicateExpr, out: &mut Vec<&'a PredicateExpr>) {
-    match predicate {
-        PredicateExpr::And(exprs) => {
-            for expr in exprs {
-                flatten_predicate_terms(expr, out);
-            }
-        }
-        _ => out.push(predicate),
-    }
-}
-
-fn predicate_term_to_condition(predicate: &PredicateExpr) -> Option<Condition> {
-    match predicate {
-        PredicateExpr::Cmp {
-            left,
-            op,
-            right: ValueRef::Literal(value),
-        } => {
-            let column = to_runtime_column(&left.column);
-            Some(match op {
-                PredicateCmpOp::Eq => Condition::Eq {
-                    column,
-                    value: value.clone(),
-                },
-                PredicateCmpOp::Ne => Condition::Ne {
-                    column,
-                    value: value.clone(),
-                },
-                PredicateCmpOp::Lt => Condition::Lt {
-                    column,
-                    value: value.clone(),
-                },
-                PredicateCmpOp::Le => Condition::Le {
-                    column,
-                    value: value.clone(),
-                },
-                PredicateCmpOp::Gt => Condition::Gt {
-                    column,
-                    value: value.clone(),
-                },
-                PredicateCmpOp::Ge => Condition::Ge {
-                    column,
-                    value: value.clone(),
-                },
-            })
-        }
-        PredicateExpr::IsNull { column } => Some(Condition::IsNull {
-            column: to_runtime_column(&column.column),
-        }),
-        PredicateExpr::IsNotNull { column } => Some(Condition::IsNotNull {
-            column: to_runtime_column(&column.column),
-        }),
-        PredicateExpr::True => None,
-        _ => None,
-    }
-}
-
-fn relation_predicate_to_conditions(predicate: &PredicateExpr) -> Option<Vec<Condition>> {
-    if matches!(predicate, PredicateExpr::True) {
-        return Some(Vec::new());
-    }
-
-    let mut terms = Vec::new();
-    flatten_predicate_terms(predicate, &mut terms);
-
-    let mut conditions = Vec::with_capacity(terms.len());
-    for term in terms {
-        let condition = predicate_term_to_condition(term)?;
-        conditions.push(condition);
-    }
-
-    Some(conditions)
-}
-
-fn extract_linear_join_info(expr: &RelExpr) -> Option<LinearJoinInfo> {
-    match expr {
-        RelExpr::TableScan { table } => Some(LinearJoinInfo {
-            base_table: *table,
-            base_scope: table.as_str().to_string(),
-            conditions: Vec::new(),
-            joins: Vec::new(),
-        }),
-        RelExpr::Filter { input, predicate } => {
-            let mut inner = extract_linear_join_info(input)?;
-            let conditions = relation_predicate_to_conditions(predicate)?;
-            inner.conditions.extend(conditions);
-            Some(inner)
-        }
-        RelExpr::Join {
-            left,
-            right,
-            on,
-            join_kind,
-        } => {
-            if !matches!(join_kind, JoinKind::Inner) {
-                return None;
-            }
-            let right_table = match right.as_ref() {
-                RelExpr::TableScan { table } => *table,
-                _ => return None,
-            };
-            let mut left_info = extract_linear_join_info(left)?;
-            let first_join = on.first()?;
-
-            let left_scope = first_join
-                .left
-                .scope
-                .clone()
-                .unwrap_or_else(|| left_info.base_scope.clone());
-            let right_scope = first_join
-                .right
-                .scope
-                .clone()
-                .unwrap_or_else(|| right_table.as_str().to_string());
-
-            left_info.joins.push(JoinSpec {
-                table: right_table,
-                alias: (right_scope != right_table.as_str()).then_some(right_scope.clone()),
-                on: Some((
-                    format!("{left_scope}.{}", first_join.left.column),
-                    format!("{right_scope}.{}", first_join.right.column),
-                )),
-            });
-            left_info.base_scope = right_scope;
-            Some(left_info)
-        }
-        _ => None,
-    }
-}
-
-fn extract_step_scan(expr: &RelExpr, predicates: &mut Vec<PredicateExpr>) -> Option<TableName> {
-    match expr {
-        RelExpr::TableScan { table } => Some(*table),
-        RelExpr::Filter { input, predicate } => {
-            predicates.push(predicate.clone());
-            extract_step_scan(input, predicates)
-        }
-        _ => None,
-    }
-}
-
-fn parse_gather_core(seed: &RelExpr, step: &RelExpr, max_depth: usize) -> Option<RuntimeCorePlan> {
-    let seed_info = extract_linear_join_info(seed)?;
-    if !seed_info.joins.is_empty() {
-        return None;
-    }
-
-    let (step_join, _step_columns) = match step {
-        RelExpr::Project { input, columns } => (input, columns),
-        _ => return None,
-    };
-    let (step_left, step_right, step_on, step_join_kind) = match step_join.as_ref() {
-        RelExpr::Join {
-            left,
-            right,
-            on,
-            join_kind,
-        } => (left, right, on, join_kind),
-        _ => return None,
-    };
-    if !matches!(step_join_kind, JoinKind::Inner) {
-        return None;
-    }
-
-    let step_hop_table = match step_right.as_ref() {
-        RelExpr::TableScan { table } => *table,
-        _ => return None,
-    };
-
-    let mut step_predicates = Vec::new();
-    let step_scan_table = extract_step_scan(step_left, &mut step_predicates)?;
-
-    let mut frontier_column: Option<String> = None;
-    let mut step_filters = Vec::new();
-    for predicate in &step_predicates {
-        let mut terms = Vec::new();
-        flatten_predicate_terms(predicate, &mut terms);
-        for term in terms {
-            match term {
-                PredicateExpr::Cmp {
-                    left,
-                    op: PredicateCmpOp::Eq,
-                    right: ValueRef::RowId(RowIdRef::Frontier),
-                } => {
-                    let candidate = to_runtime_column(&left.column);
-                    if let Some(existing) = &frontier_column
-                        && existing != &candidate
-                    {
-                        return None;
-                    }
-                    frontier_column = Some(candidate);
-                }
-                _ => {
-                    let condition = predicate_term_to_condition(term)?;
-                    step_filters.push(condition);
-                }
-            }
-        }
-    }
-
-    let inner_column = frontier_column?;
-    let first_join = step_on.first()?;
-    let left_scope = first_join
-        .left
-        .scope
-        .clone()
-        .unwrap_or_else(|| step_scan_table.as_str().to_string());
-    let right_scope = first_join
-        .right
-        .scope
-        .clone()
-        .unwrap_or_else(|| step_hop_table.as_str().to_string());
-
-    let recursive = RecursiveSpec {
-        table: step_scan_table,
-        inner_column,
-        outer_column: "_id".to_string(),
-        select_columns: None,
-        filters: step_filters,
-        joins: vec![JoinSpec {
-            table: step_hop_table,
-            alias: (right_scope != step_hop_table.as_str()).then_some(right_scope.clone()),
-            on: Some((
-                format!("{left_scope}.{}", first_join.left.column),
-                format!("{right_scope}.{}", first_join.right.column),
-            )),
-        }],
-        result_element_index: Some(1),
-        hop: None,
-        max_depth,
-    };
-
-    Some(RuntimeCorePlan {
-        table: seed_info.base_table,
-        conditions: seed_info.conditions,
-        joins: Vec::new(),
-        result_element_index: None,
-        recursive: Some(recursive),
-    })
-}
-
-fn parse_gather_join_info(expr: &RelExpr) -> Option<GatherJoinInfo> {
-    match expr {
-        RelExpr::Gather {
-            seed,
-            step,
-            max_depth,
-            ..
-        } => {
-            let plan = parse_gather_core(seed, step, *max_depth)?;
-            Some(GatherJoinInfo {
-                current_scope: plan.table.as_str().to_string(),
-                plan,
-            })
-        }
-        RelExpr::Filter { input, predicate } => {
-            let mut inner = parse_gather_join_info(input)?;
-            let mut conditions = relation_predicate_to_conditions(predicate)?;
-            inner.plan.conditions.append(&mut conditions);
-            Some(inner)
-        }
-        RelExpr::Join {
-            left,
-            right,
-            on,
-            join_kind,
-        } => {
-            if !matches!(join_kind, JoinKind::Inner) {
-                return None;
-            }
-            let mut left_info = parse_gather_join_info(left)?;
-            let right_table = match right.as_ref() {
-                RelExpr::TableScan { table } => *table,
-                _ => return None,
-            };
-            let first_join = on.first()?;
-            let left_scope = first_join
-                .left
-                .scope
-                .clone()
-                .unwrap_or_else(|| left_info.current_scope.clone());
-            let right_scope = first_join
-                .right
-                .scope
-                .clone()
-                .unwrap_or_else(|| right_table.as_str().to_string());
-
-            left_info.plan.joins.push(JoinSpec {
-                table: right_table,
-                alias: (right_scope != right_table.as_str()).then_some(right_scope.clone()),
-                on: Some((
-                    format!("{left_scope}.{}", first_join.left.column),
-                    format!("{right_scope}.{}", first_join.right.column),
-                )),
-            });
-            left_info.current_scope = right_scope;
-            Some(left_info)
-        }
-        _ => None,
-    }
-}
-
-fn parse_runtime_core_plan(core: &RelExpr) -> Option<RuntimeCorePlan> {
-    match core {
-        RelExpr::Gather {
-            seed,
-            step,
-            max_depth,
-            ..
-        } => parse_gather_core(seed, step, *max_depth),
-        RelExpr::Project { input, .. } => {
-            if let Some(mut gather_info) = parse_gather_join_info(input) {
-                if !gather_info.plan.joins.is_empty() {
-                    gather_info.plan.result_element_index = Some(gather_info.plan.joins.len());
-                }
-                return Some(gather_info.plan);
-            }
-
-            let linear = extract_linear_join_info(input)?;
-            Some(RuntimeCorePlan {
-                table: linear.base_table,
-                conditions: linear.conditions,
-                joins: linear.joins.clone(),
-                result_element_index: Some(linear.joins.len()),
-                recursive: None,
-            })
-        }
-        _ => {
-            if let Some(gather_info) = parse_gather_join_info(core) {
-                return Some(gather_info.plan);
-            }
-
-            let linear = extract_linear_join_info(core)?;
-            Some(RuntimeCorePlan {
-                table: linear.base_table,
-                conditions: linear.conditions,
-                joins: linear.joins,
-                result_element_index: None,
-                recursive: None,
-            })
-        }
-    }
-}
-
-fn unwrap_query_envelope(expr: &RelExpr) -> QueryEnvelope<'_> {
-    let mut current = expr;
-    let mut order_by = Vec::new();
-    let mut offset = 0;
-    let mut limit = None;
-
-    loop {
-        match current {
-            RelExpr::OrderBy { input, terms } => {
-                if order_by.is_empty() {
-                    order_by = terms
-                        .iter()
-                        .map(|term| {
-                            (
-                                term.column.column.clone(),
-                                match term.direction {
-                                    OrderDirection::Asc => SortDirection::Ascending,
-                                    OrderDirection::Desc => SortDirection::Descending,
-                                },
-                            )
-                        })
-                        .collect();
-                }
-                current = input;
-            }
-            RelExpr::Offset { input, offset: n } => {
-                offset = *n;
-                current = input;
-            }
-            RelExpr::Limit { input, limit: n } => {
-                limit = Some(*n);
-                current = input;
-            }
-            _ => {
-                return QueryEnvelope {
-                    core: current,
-                    order_by,
-                    offset,
-                    limit,
-                };
-            }
-        }
-    }
-}
-
-fn build_execution_query_from_relation(query: &Query, relation: &RelExpr) -> Option<Query> {
-    let envelope = unwrap_query_envelope(relation);
-    let core_plan = parse_runtime_core_plan(envelope.core)?;
-
-    let mut lowered = query.clone();
-    lowered.table = core_plan.table;
-    lowered.alias = None;
-    lowered.disjuncts = vec![Conjunction {
-        conditions: core_plan.conditions,
-    }];
-    lowered.joins = core_plan.joins;
-    lowered.order_by = envelope.order_by;
-    lowered.offset = envelope.offset;
-    lowered.limit = envelope.limit;
-    lowered.result_element_index = core_plan.result_element_index;
-    lowered.recursive = core_plan.recursive;
-    lowered.select_columns = None;
-    lowered.relation_ir = None;
-    Some(lowered)
-}
-
-fn build_execution_query_from_relation_ir(query: &Query) -> Option<Query> {
-    let relation = query.relation_ir.as_ref()?;
-    build_execution_query_from_relation(query, relation)
 }
 
 impl QueryGraph {
@@ -698,7 +243,7 @@ impl QueryGraph {
         let mut template = Query::new("__relation_ir");
         template.branches = branches.to_vec();
 
-        let query = build_execution_query_from_relation(&template, relation)?;
+        let query = lower_relation_to_execution_query(&template, relation)?;
         Self::compile_with_session(&query, schema, session)
     }
 
@@ -712,7 +257,7 @@ impl QueryGraph {
         session: Option<Session>,
     ) -> Option<Self> {
         let relation_query = if query.has_relation_ir() {
-            Some(build_execution_query_from_relation_ir(query)?)
+            Some(lower_query_relation_ir(query)?)
         } else {
             None
         };
@@ -939,7 +484,7 @@ impl QueryGraph {
         schema_context: &SchemaContext,
     ) -> Option<Self> {
         let relation_query = if query.has_relation_ir() {
-            Some(build_execution_query_from_relation_ir(query)?)
+            Some(lower_query_relation_ir(query)?)
         } else {
             None
         };
@@ -2462,8 +2007,8 @@ mod tests {
     use super::*;
     use crate::query_manager::query::QueryBuilder;
     use crate::query_manager::relation_ir::{
-        ColumnRef, JoinCondition, KeyRef, PredicateCmpOp, PredicateExpr, ProjectColumn,
-        ProjectExpr, RelExpr, RowIdRef, ValueRef,
+        ColumnRef, JoinCondition, KeyRef, OrderByExpr, OrderDirection, PredicateCmpOp,
+        PredicateExpr, ProjectColumn, ProjectExpr, RelExpr, RowIdRef, ValueRef,
     };
     use crate::query_manager::types::{ColumnDescriptor, ColumnType, RowDescriptor, Value};
 
@@ -3012,6 +2557,75 @@ mod tests {
         assert!(
             has_recursive_relation_node(&graph),
             "Recursive join-projection queries should compile to RecursiveRelationNode"
+        );
+    }
+
+    #[test]
+    fn compile_query_with_relation_ir_project_join_order_limit_shape() {
+        let schema = recursive_hop_schema();
+        let relation = RelExpr::Limit {
+            input: Box::new(RelExpr::OrderBy {
+                input: Box::new(RelExpr::Project {
+                    input: Box::new(RelExpr::Join {
+                        left: Box::new(RelExpr::Filter {
+                            input: Box::new(RelExpr::TableScan {
+                                table: TableName::new("team_edges"),
+                            }),
+                            predicate: PredicateExpr::Cmp {
+                                left: ColumnRef::scoped("team_edges", "child_team"),
+                                op: PredicateCmpOp::Eq,
+                                right: ValueRef::Literal(Value::Integer(7)),
+                            },
+                        }),
+                        right: Box::new(RelExpr::TableScan {
+                            table: TableName::new("teams"),
+                        }),
+                        on: vec![JoinCondition {
+                            left: ColumnRef::scoped("team_edges", "parent_team"),
+                            right: ColumnRef::scoped("__hop_0", "id"),
+                        }],
+                        join_kind: crate::query_manager::relation_ir::JoinKind::Inner,
+                    }),
+                    columns: vec![ProjectColumn {
+                        alias: "id".to_string(),
+                        expr: ProjectExpr::Column(ColumnRef::scoped("__hop_0", "id")),
+                    }],
+                }),
+                terms: vec![OrderByExpr {
+                    column: ColumnRef::unscoped("name"),
+                    direction: OrderDirection::Desc,
+                }],
+            }),
+            limit: 5,
+        };
+
+        let mut query = QueryBuilder::new("placeholder").branch("main").build();
+        query.relation_ir = Some(relation);
+        query.recursive = None;
+        query.joins.clear();
+        query.select_columns = None;
+
+        let graph = QueryGraph::compile(&query, &schema).expect("Graph should compile");
+        assert!(
+            graph
+                .nodes
+                .iter()
+                .any(|ctx| matches!(ctx.node, GraphNode::Join(_))),
+            "relation IR join shape should compile to JoinNode",
+        );
+        assert!(
+            graph
+                .nodes
+                .iter()
+                .any(|ctx| matches!(ctx.node, GraphNode::Sort(_))),
+            "relation IR order by should compile to SortNode",
+        );
+        assert!(
+            graph
+                .nodes
+                .iter()
+                .any(|ctx| matches!(ctx.node, GraphNode::LimitOffset(_))),
+            "relation IR limit should compile to LimitOffsetNode",
         );
     }
